reports: send add_path valid links in html_links
As of commit 1b90f02, we no longer just pass through invalid file names,
because we try to find out where they're located. This caused issues
with this line of code, which tried to first get a fully qualified path,
and then sprintf the icon name into place. Do it the sane way around
instead.

Change-Id: I30d17f7fceb3db3c76a3a6cf848827fc2030d5f8
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>