Kohana: Remove ReflectionClass usage in (some parts of) Kohana
ReflectionClass is really nice. It does debugging much simpler, and can
be really useful for decorating Exceptions and such. But it isn't meant
to be used for real time digging and analysis of controller classes.
For that usage, variable variables, and calling named methods is much
better to use. Like $controller->$method($args).

In fact, lot of code was to analyze with ReflectionClass if a method
existed, or if the __call method should be used, when it's in fact is
just to call the method and let the PHP backend do its work.

So just drop the usage of ReflectionClass from Kohana, and the code is
much more readable.

The only "drawback" is that we don't know explicitly if the class is
abstract, but in fact, if it is abstract, we can't instance it, and
thus we get an error. But it probably needs to be handled a little bit
nicer (go to index.php/template, which is the abstract root of many
controllers; the Template_Controller)

Change-Id: I69895cb7af9a5f86c91aa86c668f745acace7a92
Signed-off-by: Max Sikstrom <msikstrom@op5.com>