Kohana: Use proper 404 page if page is not found
By some reason, in the ninja controller, we trapped all unknown methods
with a __call to display a black text string on a white background
saying that the page doesn't exist, instead of showing a nice page,
which is default, if the page doesn't exist.

In fact, we still show the nice page if we go to an unknown controller,
but if the controller is known but the action is unknown, we show the
ugly one. And that is made explicitly by adding a strange __call method
in the Ninja_Controller, instead of using the one that makes sense in
the root class Controller.

So just remove the __call from Ninja_Controller, and everything looks
much nicer :)

Change-Id: Ib80cd63ef3369c1bf4e0b76942e60cca88d271e8
Signed-off-by: Max Sikstrom <msikstrom@op5.com>