Kohana: Convert PHP Errors to exceptions
PHP errors and exceptions were handled by the same method in Kohana,
but everything was a little bit different.

PHP errors is practically an exception, but genereted implicitly. To
simplify life, convert the php error to an exception, and enter the
exception handling scheme recuces the complexity of the exception
handling.

Change-Id: I702a92ad11b83e151f4973863b8253061e662a71
Signed-off-by: Max Sikstrom <msikstrom@op5.com>