Kohana: Convert backtrace view to a view
A view is a file, that given a couple of parameters outputs the
information user readable.

The Kohana class is the glue that holds everything together.

The Kohana::backtrace method took one argument, a stack trace, and
returned a user readable page describing the stack trace, which per
definition would be a view.

So lets simply convert the Kohana::backtrace to a view file, and treat
it as a view.

One might think that it's not a good idea to have views in the error
handler if something went wrong. But views itself is one of the most
isolated pieces of code in kohana, so that should be ok. Better to use
the generic view than hand-crafting require()-statements that itself
might be incorrect.

Change-Id: Id8ad795655b805de71d6731bda0a5c7862a6b710
Signed-off-by: Max Sikstrom <msikstrom@op5.com>