Tests: Mock Session class for unit tests
There is no way in Kohana to disable session cookies using the Session
driver. That means, we should mock it up for testing

Easiest is just to make sure the mock-driver is in place before loading
kohana, which means it simply just won't autoload, because it is
believed already to be loaded.

The mock-driver doesn't do anything, but return the default value every
time a variable is meant to be recieved.

Change-Id: Ia375dc16f4d577961462349a76233ee363902e3c
Signed-off-by: Max Sikstrom <msikstrom@op5.com>