sysinfo: make get_merlin_usage more failsafe
Don't trust the kvvec to be a valid kvvec; ignore parts without an =

Ignoring parts would just end up finding fewer peers, which would
possible give a user an extra peer slot instead of adding problems. If
this is an error, the user already have more trouble with the system
anyway.

(This commit is imported from op5lib)

Change-Id: I6eea4039e46bb7275f2b32a7812257ce2fa63743
Signed-off-by: Max Sikstrom <msikstrom@op5.com>