jq-filterable: Don't expect id to contain [
The name of a select (when sending to PHP) should contain [], so PHP
interprets the name as an array. The id however shouldn't have it,
since the css selectors will be messed up.

Therefore, replacing [ with _tmp[ isn't a good way of making the
non-selected clone select having a unique id. So simply append _tmp to
the name instead.

Resolves an issue masked by #9029, resulting that objects=null in ajax
request

Change-Id: Id0ab775d932e57474968c8329b0f87be19c11511
Signed-off-by: Max Sikstrom <msikstrom@op5.com>