recurring_downtime: Restore original name for downtime_type
The SQL structure depends on the downtime_type name, as does the
controller, so let's just restore it. It was called downtime_type up
until 5c975b6, at which point it was accidentally renamed, due to
recurring_downtime/setup.php having an almost but not quite identical
structure as objselector.php.

As should be apparent from this mistake, it would be better to merge the
two, but the differences between them are still sufficient to make me
punt it for now and leave it to be taken care of on master.

This means I have to change the selector in common.js from
[name=report_type] to #report_type, so I can create a
[name=downtime_type]#report_type for use by the generic code.

While doing this, clean up some related, unused code (#sel_*_type
doesn't exist, code that's commented out isn't used, slightly fewer
recurring_downtime.js special cases).

This resolves bug #9029.

Change-Id: Ib59f0e80253229ab812993468e561ca4b04f6cab
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>