ORM generator: Reorganize files
We need to refactor the generator to have driver specific pools, not
nessecary sets, so we can move the database access to the pool instead
of set.

That would make it possible to have have the object write things back
to the database without accessing sets in the future. Just a first step
to add write support in the ORM

Change-Id: Ib14f4bfac2c6865cbc70df1bdc9a74c912d2ad1f
Signed-off-by: Max Sikstrom <msikstrom@op5.com>