ORM generators: Initial write support
Since it would be nice to have write support, we should probably have
it, and thus, implement that...

Makes it possible to use the same models for reading and writing.
First, it only implements write support to saved filters as PoC, but
would be possible to do on all database tables later.

To create:
$obj = new SomeStuff_Model();
$obj->set_field(data);
$obj->set_another_field(data);
$obj->save();
echo $obj->get_id();

To update, just set the fields and save.

Change-Id: I1ed3bebdc3ce802d98a2f8c8d6d6060c3e49f2fc
Signed-off-by: Max Sikstrom <msikstrom@op5.com>