ListView: JS error in custom columns error handler
Creating a custom column with errors in the naming (which means before
the = sign), js tried to call an undefined method, and crashed.

The reason for this was when sorting on custom columnns where added,
the generator methods was wrapped in an object, and augmented with
sorting information. This was missing in one of the error handlers.

Rsolves: #8853

Change-Id: I636b1257c068be80fbed9d59202737f9e928a81c
Signed-off-by: Max Sikstrom <msikstrom@op5.com>