sql: Add default log filters
This change adds a handful of default log filters for the monitor user.

They should serve as good examples for what is possible to do with the
logfilter functionality, while at the same time providing functionality
that could be copy-pasted to be useful for a majority of users.

They're only defined for the monitor user as to not clutter up the Saved
Filters list for all users, since we don't have filter hierarchies
implemented yet (#7165).

This fixes #8880.

Change-Id: I41b550c5a37028250f2865bf4515ec30c3d55fae
Signed-off-by: Anton Lofgren <alofgren@op5.com>