cli: Fix PHP warning in nacoma when renaming objects
If a saved (summary) report contains the object selection "All objects",
it will not be represented as a list of all the objects in the system at
the time, but as a magical string token (which happens to be "*"). This
wasn't handled in the rename/deletion hook, so instead, the hook spat
warnings at the user. This is ugly.

Change-Id: I3cc2424d612bd95e3d5765559aead9ecaadc5c32
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>