MayI: Implement actor interface
An actor is an object (like the user, like the web server) providing
information to the constraint interface, as source set for the
contraints.

The actor is implemented as an op5MayI_Actor interface, only expecting
the method getActorInfo to exist, to export the information about the
actor.

Part of #9134

(This commit is imported from op5lib)

Change-Id: I872b191744fecfc4d9f71296ef0cf38dec67f804
Signed-off-by: Max Sikstrom <msikstrom@op5.com>