ls_filter: Make visuals for time filter semantically correct
If the user would use the filter builder to fetch everything that was
for example checked within the last hour it would be logical to set
something like: last_check less than 1 hour ago.
This would however give the user everything that was checked earlier
than the last hour since the column takes timestamps as values. Less
than would in that case return everything with a lower timestamp than 1
hour ago.

Instead we want the user to be able to build the filter like this:
last_check later than 1 hour ago. This will yield the wanted results.

Change the option texts to reflect how a user would ask a question about
time.

This fixes #9126

Change-Id: I3ec45bac001bf6c066787585cc6457ded3c3eb50
Reported-by: Ola Sandström <osandstrom@op5.com>
Signed-off-by: Henrik Smederod <hsmederod@op5.com>