template_menu: Always print icons in span-tags, even external links'
The HTTP API icon had an ugly offset because it was treated in a
different way because of the third parameter in The Menu Array.

I'm leaving the third clause untouched since I couldn't find anything
that uses it.

Also cleaned up some stuff (but not the duplicated code, since we'll
remake this very soon?).

Change-Id: I411ae8a94655936dd9e46b02060c2a51f5eb9288
Signed-off-by: Carl Helmertz <chelmertz@op5.com>