Remove unused code
Ninja is a nest of old hacks, and the old Ninja_Controller isn't better.

Remove lot of unused code, to simplify the controller.

There are three major things removed here:

 - Generic sort column handling, which is sorting of the old monitor
6.0 and earlier lists (pre-list view lists)

 - Help link, which was a test concept made back in 6.0 to have links
to help pages which should contain information of from where the link
was sent. But since the url is never set, and the other end never was
implemented, or even thought about, this can safely be removed. If
needed, we simply re-implement it in a nicer way later.

 - Removal of some old pagination-get-variable-mangling which is just
strange and not too generic for the generic place it's placed in.

Change-Id: I8d5c4c7a69418ee691f0567ca3b6ba43f2386ae5
Signed-off-by: Max Sikstrom <msikstrom@op5.com>