Ninja controller refactor: break out module specific code to modules
The ninja controller holds some ninja-specific features like loading
the main template, adding an infrastructure for notifications, and
simliar. But it also held code like check for monitoring features,
which should be a part of the monitoring module, and some
nacoma-specific tests for autoscan results.

The modules have the ability to add notifications from hooks, so it can
hook into the lifecycle of the page load, and then have access to do
these kinds of tests after the controller is executed.

Breaking out the code to hooks makes the Ninja class cleaner, and the
system more livestatus/naemon/nagios/whatever-dependent in its core,
and the monitoring is moved to a module.

Also remove xlinks-array, which I can't find any refernces to at all in
our repositories, and some variables that wasn't ever used.

Change-Id: I301239c359a5c21932c9076ab460b27f3ff125aa
Signed-off-by: Max Sikstrom <msikstrom@op5.com>