Tests: improve tests of GET variable auth
Since we haven't defined exactly how GET variable authentication should
be handled, we need to define it, so lets start with some tests.

The behaviour is as follows ATM:
 - When logged in, the session is used
 - If not logged in, username, password and optional auth_method is
defined, try to log in, and update session
 - After logging in with get auth, the user stays logged in

This behaviour might or might not be expected, but at least it's now
defined using tests, and can be refactored, or formally updated.

Change-Id: I415a185b04ba86cfeeb20e84f9f8c6b29d3b82c5
Signed-off-by: Max Sikstrom <msikstrom@op5.com>