Clean up cli tests
There were some unnessecary tests for if the user was using cli.

If the user is accessing something other than the cli controller, he
should probably manage to waste a couple of extra CPU clock cycles to
load the menu and addons, which is quite fast, since it's loaded each
and every page load anyway.

Also, we have removed user roles since long, so don't test for that in
the cli controller

Change-Id: I06b6f5fa8b838aecfb54f2569427aab1acce620d
Signed-off-by: Max Sikstrom <msikstrom@op5.com>