Add event system.render, and move rendering to that
Rendering was added to the system.post_controller event in classic
kohana style, since Kohana used output buffering. That meant that
headers could be sent in the Kohana shutdown method, just before the
output buffer was dumped to the user.

Since we have removed the output buffering, since that's really nasty,
the output was started in system.post_controller instead of
dKohana::shutdown, and thus, mangling HTTP headers in the
post_controller would be impossible.

If rendering is slow, which would be quite acceptable, it would be nice
for the authorization and session handling to be shut down before
rendering, but after the controller is executed. Thus we need to
guarantee that the output hasn't started at system.post_controller, and
it just makes sense to add rendering in system.render after
system.post_controller in the current structure.

Change-Id: Ic9939a20831acd9bc19edcba0b234db1b9c13173
Signed-off-by: Max Sikstrom <msikstrom@op5.com>