Auth: Add write_close, to write back and close session + Test
Since there are many reasons to make the authentication interface
volatile, and close the session, we should support that.

 - To temporarily gain access, for example for cron jobs, tests

 - Rendering shouldn't change the session anyway, and might take time,
so make it possible to unblock the session.

To implement this method, we need tests, and since op5auth were somwhat
untested (the drivers are quite well tested though), add some basic
tests to that class too.

(This commit is imported from op5lib)

Change-Id: Idead24f214606d5566276a5dea3cba526788d849
Signed-off-by: Max Sikstrom <msikstrom@op5.com>