command: Always validate input timestamps
When trying to schedule a downtime for a service, we didn't validate the
dates prior to checking if there downtime is a retrospecively added one,
thus we failed hard to generate the proper SQL, thus stack tracing.

This fixes the service part of #7272 and also adds similar validation of
timestamps to other commands using it as input, such as downtime for
objects in host- or servicegroups.

Besides that, fix the return type of _validate_dates() by introducing an
error message reference as a parameter, which is the smart way to do
stuff. Yay.

Change-Id: I0a9310fe0dba0a81d2f2261217025cfffba682fb
Signed-off-by: Carl Helmertz <chelmertz@op5.com>