Run auth in system.pre_controller instead of system.ready
Authentication should be executed before the controller constructor,
but after the system is set up. Thus, it should be executed in one of
the first events.

The system.ready event is executed to make the system ready, which
should probably be fine. But because the system.ready event treated as
last-minute setup for the system (like updating the mayi actors and
contstraints), it shouldn't hold any authentication logic, so we don't
really depend on the order of the callbacks in the events.

But since the controller execution has two steps: controller
constructor and controller execution, and since there are events named
system.post_controller and system.post_controller_construct, and also
system.pre_controller, it would be easy to assume that there would be a
system.pre_controller_constructor too, like the order:

system.pre_controller_constructor
$controller = new $controller(); // controller constructor
system.post_controller_constructor

system.pre_controller
$controller->$method();
system.post_controller

but no... that's not the Kohana way. This is the Kohana way:

system.pre_controller
$controller = new $controller();
system.post_controller_constructor
$controller->methods();
system.post_controller

It's not incorrect in any way, but it's... well... not as I would have
named it.

So, simply do authentication in the system.pre_controller, which is
practically where we did it earlier when integrated in the
Authenticated_Contorller::__construct()

Change-Id: I9eb8826b5936b63a668038e93749e8ed1312dc91
Signed-off-by: Max Sikstrom <msikstrom@op5.com>