Default controller: Minor cleanup
This code wasn't used anywhere, so remove it.

The no_objects method didn't do anything that is useful at all with the
current auth, and the view didn't even work.

The GET auth is already moved from the authenticated_controller to a
hook, and was never handled at all in the do_login page.

So just throw everything out, be happy, and dance around a bit

Change-Id: I3124609b5b88081444cb394a0f180b880d107e18
Signed-off-by: Max Sikstrom <msikstrom@op5.com>