Refactor login page: Remove ninja_auth, use query string for return URI
The ninja_auth did handle lot of strange stuff for login, such as max
invalid login attempts before locking the user out. (Which was session
based, so the user just needed to go to index.php/default/logout and
start over). And also locking the user out if it didn't have access to
hosts or services (but didn't bother to look at for example log
messages or reports or anything else).

So just throw ninja_auth out of ninja, so we only have the Auth library
and op5Auth library left, beside the default controller that actually
handles the authentication interface to the user, beside the hook
handling GET-based auth and CLI-auth.

Resolves #9245

Change-Id: I21ced247909ef9c96e82fa8fcc53e7dbd572c779
Signed-off-by: Max Sikstrom <msikstrom@op5.com>