Refactor default controller: Use a single login action
Since login is done as POST, using a POST request should attempt login,
and GET should just display the login form

If the user is logged in in any case, we should just redirect to the
requested or default url. Otherwise we should display a login form. No
need for different actions for login.

This will however break the saved URL:s for the login page. So add a
backward-compatibility wrapper.

Change-Id: I0b050f5012d85fa91f78f5a77ffe0e57e8a5e413
Signed-off-by: Max Sikstrom <msikstrom@op5.com>