js: Stay away from alert
According to GDB, when phantomjs suddenly hangs, it's because of Qt is
trying to display an alert, and poltergeist didn't expect it - which
is funny, considering that poltergeist doesn't support not accepting
alerts. Either way, this makes me think alerts are evil.

In addition, alerts are ugly, and $.notify is much cooler. Let's use
that instead.

Change-Id: I3bab3fddd4d669a18ae30d7e33cf7a71cf1ac73e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>