ldap: Fetch diagnostic error message if available
Whenever we throw an ldap error check if there is additional information
available for that error and include that in the error message.

This can for example fetch this message when the users certificate is
not in order:
Error Binding to LDAP: TLS error -8179:Peer's Certificate issuer is not
recognized.

(This commit is imported from op5lib)

Change-Id: I1046f71e01b7faf72b08b7e63194e09e29a49847
Kudos-to: Mikael Falkvidd <mfalkvidd@op5.com>
Signed-off-by: Henrik Smederod <henrik@smederod.com>