external_widget: Search additional places for widget
If one wants to view a certain instance of external widget the
instance_id can be set in the url to fetch the correct one. It is
however not certain that the instance we are looking for exists on the
page where we call the widget (external_widget/show_widget) but it
should exist on the page tac/index.

Check if it exists on tac/index before returning false.

This fixes #9296

Change-Id: If99c36c00a46f66525b4d68d4a9f25448850fa91
Signed-off-by: Henrik Smederod <henrik@smederod.com>