Remove old apache login "module" ninja_login.php
This has probably not been used for many years. Nothing in the
manual or KB mentions this feature. We have a better way to
support apache authentication, see "Enable Apache Auth in op5
Monitor", https://kb.op5.com/x/g4EK
And it sets cookes that are not httponly and secure, which was
detected by Hexbit in the 7.0 pentest.

Change-Id: I2256d545dc6ee6b06c9fe32e91d4189f9e440c38