Treat system.ready event as part of loading Kohana
The event system.ready is treated as an event to load custom parts into
a loaded Kohana at startup, like authorization, mangling the state of
Kohana itself.

When loading index.php with SKIP_KOHANA set to true, it is expected to
have access to a working and complete Kohana as loaded in ninja, with
access to all models and so, but nothing should really be executed.
That means all initialization, also the parts related to modules,
should be executed.

At the moment, system.ready loads only MayI-modules, which will with
this patch be available also to the API, so authorizaiont implicitly is
enabled everywhere ORM is available.

Change-Id: I521bed1a0adb0da28224472752a0fc035987a4db
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>