db_migrations: Keep schedules, usernames around after upgrade
The old code didn't care enough to preserve the username, which causes
limited users to loose permissions to the report, or to preserve
schedules, which means that when scheduled reports run, either no
reports will be mailed out at all, or the wrong ones will.

This fixes bug #9319 and #9320. This does not fix systems that are
already broken - the above-mentioned bugs contains fixes for both the
problems, though.

Change-Id: I4dac73f65fbc5821a90c4f38a84ef30e52cb7d8a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>