Remove Hypermap
Since it was retired in monitor 7.0, and this is monitor 7.1, we
doesn't support it anymore. At the moment, everything about
authorization and licensing is rewritten, it needs to be rewritten for
hypermap too.

The hypermap isn't well tested, and we have no simple means to verify
the authorization integration, and the time should be focused upon what
we support, we remove the hypermap all togheter.

This removes traces of #7689

Change-Id: Id87a126ac1ee244aa949271b73037fc4cdaa6ead
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>