reports: Restore selection adjusting
At some point during filterselect conversion, we managed to lose the
code that hid the irrelevant state options from the report setup -
i.e. the service-specific options for host reports and vice
versa. This merely restores this code, except (spoiler alert!) does so
in a way that op5reports will start to exploit as soon as a bunch of
other patches have been written.

Change-Id: I2617223c726dac905b4d5e2025bdf4d1a8906792
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>