db_migrations: Don't upgrade same schedule twice
mfalkvidd managed to trigger a case where he had one avail report (id
1 in the avail_config) and 3 sla reports (id 1, 2, and 3 in the
sla_config). When migrating to the new table format, the avail report
was given id 1, and the sla reports were given id 2, 3 and 4,
respectively. This meant the migration script did what basically
amounted to:

 update report_id = 1 where report_id = 1 and type is avail
 update report_id = 2 where report_id = 1 and type is sla
 update report_id = 3 where report_id = 2 and type is sla
 update report_id = 4 where report_id = 3 and type is sla

The avail bit works great. The sla bit, not so much.

To fix this case, we now keep track of the reports that matched the
last expression time, so we can avoid updating those again.

Change-Id: If8a586440c379393a63b8269f0566c21006f265c
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>