reports: Remove alert_types
alert_types used to let you choose between host states or service
states. It was defined for summary reports, which also had a host
state selector and a service state selector.

In other words, if you only want host states, don't check service
states - you don't need an additional dropdown for selecting the same
thing again.

This resolves bug #8143.

In the HTTP Event API, alert_types was exposed, and thus it would be
an API break to remove it completely. We still handle the value, if
provided (like we must do for saved reports), so it's not an actual
API break - although the documentation for the property has vanished,
as we have no good way to include documentation for deprecated
properties.

Change-Id: Ia8695746faaf538efb2100dc15ed687010031dc0
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>