reports: Replace includesoftstates with state_types
Summary reports had a dropdown where you get to select hard states
only, soft states only, and both. Status reports had a checkbox to
let you choose between hard states only and both.

This replaces status reports' checkbox with a copy of the summary
reports' dropdown. This resolves bug #7421. It's sort-of a start for
fixing bug #7050, depending on if you count the comments in that bug
or not.

Change-Id: I9ebccb45852d65a53a348bfec2ef36095d1eff2d
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>