reports: Make report state filtering consistent, compatible
In summary reports, "states to 𝐢𝐧clude" was a dropdown that contained
a bunch of predefined bitmaps of states. In alert history, it was a
bunch of checkboxes. In avail/sla, states to 𝐞𝐱clude was
checkboxes. And the states you excluded from avail/sla didn't carry
over to embedded alert histories/summary reports. In short, things
were a bit of a mess.

Fix by always having the avail/sla style list of states to map and
using the magical "exclude" state for excluding a particular state.
However, represent these with alert history style "states to include"
checkboxes. Because selecting something you want and deselecting
something you don't want makes UX sense.

This sudden outburst of unparalleled consistency resolves bug #7050.

And, because I'm unable to do one thing at a time, and my commits
sometimes end up being a bit of a (fairly big) mess as a result, I
also tried to make the report forms look more consistent (add
helptexts, reuse code, make form positions more consistent) while I
was making the behavior consistent, and suddenly, too much code was
changed at once to make separate commits of that. I suck.

For avail/sla reports, when you uncheck a state, you can choose which
state to map the excluded state to. This means you can decide that
warning values are OK as far as your SLA numbers are concerned, and
thus map them that way as far as percentages are concerned. Or, for
that matter, you can choose to render all pending as OK - apparently,
this was how people liked to implicitly do it before monitor 6, when I
removed the possibility to configure the pre-first-check state.

And because all states are mapped individually, you could map unknown
to critical, but warning to OK.

This resolves another (the third, to be exact) usecase in bug #8159.

These mappings only affect the SLA/avail reports - unless you map
something to "hidden", it will be displayed as the original state any
embedded reports. I haven't yet decided whether it really would be
appropriate to have an embedded alert history render warnings as OK
when you map warnings to OK (after all, there will probably be a green
shield next to the check output "WARNING:" which would be weird), or
whether it would be confusing to have warning alerts in a report that
shows 0% warning. So, I'm going with the easiest solution.

Also rename "pending" to "undetermined". This resolves bug #6670. And
call "excluded" "hidden", now that users will see that string, because
that seems like a better description of what it does.

Change-Id: I123b1070c4486047ecc653823cde351463f2aaa6
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>