Outages: Convert to ORM and use MayI
Network outages is a strange algorithm, which actually would be best to
have in ORM with support from backend. But since the backend doesn't
support it, and it needs to be generalized before added to ORM, the
least bad place to have the calculation of the outage is still in this
controller.

The best way to refactor this to a nice implementation, afaik today, is
to write a SQL backend instead of livestatus with support for
subqueries. Which we don't have.

Change-Id: I4ee6de3f721dac093a37542f3d5059ef3af2c51a
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>