recurring_downtime: Clean up JS
First of all, adding an alert('Po is best teletubby') on top of
recurring_downtime.js alerted twice. That's not good.

Secondly, we ran check_form_values() which is report specific and is
irrelevant at best, in this context.

Thirdly, we didn't use $.notify() but instead used something that
required us to move (!) the user's viewport to the top of the page.
That's awfully douchy of us.

Lastly, Enter is nott a name (maybe your thinking of Ender?), so lets'
decapitate it (this is how it feels to read our English).

Change-Id: I521fbe8b7db13a26a1f5b0928e4d44696afd97f1
Signed-off-by: Carl Helmertz <chelmertz@op5.com>