Only require nacoma from tests
Ninja and nacoma shouldn't depend on each other, and nacoma should
install itself in ninja, since ninja should be treated as a framework.

But we have tests that depneds on nacoma, and therefore we need nacoma
to be installed during testing. Therefore, add nacoma to the
ninja-tests package, but not in ninja

Change-Id: I4140a600c4dd60bb13cb5bdc1d1762f8247fdd8c
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>