reports.sass: Improve the look of report forms slightly
The rule for "table.setup-tbl input" was setting up padding between
rows of options - but it didn't do so when the row consisted
exclusively of dropdowns.

The skin selector is one line, while it's always next to the
description entry. We should top-align it, so all the text is as
snuggly with regards to their labels.

Change-Id: I5e91670eae565364012c8eb453d170f1945190c5
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>