reports js: Fix two minor saving bugs
When a user tries to save an already generated report, we check if
it's valid. We know it's valid, because the user managed to configure
and render the report, and there are no user-editable values between
what was rendered and what is saved. So we don't need to check
anything. (what about users fiddling with their HTML tags manually?
Well, they might as well fiddle with the script - this is still all
frontend validation)

When a user saves a report, we only autohide the report save form if
the user edited an existing report, not if it's a new one. Why‽

Change-Id: I96282173c7823a92b7739ae459929a29ce6108e9
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>