report_options: Add check to verify option integrity
This should provide me, or another developer, with better error
messages in case we fuck up during development.

Not using assert, because I need PHP 5.4 to be able to add a message
to the assert failure. Fine, have an un-subtyped exception - if you
ever find yourself wanting to catch this exception, you're doing it
wrong, so not creating a subtype is acceptable.

Change-Id: Ibac9987e405ac1efb80da6b7fc02384d3dd0afe0
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>