json: Add json pretty-printer
If you showed me this code, I'd have have rejected it, because it's
obviously silly (who would write a json encoder without a
pretty-printing option?) and it's slow (because it re-does work it's
already done over and over).

For the first point, the answer is "PHP, that's who" - the built-in
pretty printer showed up in 5.4, we're on 5.3.

For the second, PHP lacks a useful array type, so the encoder relies
on magic heuristics for determining if it's an associative array (json
object) or a non-associative array (json array). I could try to
re-implement these heuristics (and likely fail, because code), or I
can re-use what the existing implementation provides by letting it
encode first and then redoing it myself, knowing that pretty-printing
should only be done for user-presented data, and user-presented data
needs to be short to be readable.

Change-Id: Ie45dfe7946e84d0a95fbfe5253578968e3c825be
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>