reports: Test that invalid options are removed, use defined exceptions
The discussion in the review request for I4d29746 put the burden of
proof on me as I wanted to claim that e.g. enums and ints are properly
validated. While validate_value() trivially does not allow those types
to be set to just anything, what if it's a value that the model
listens to, but the subreport type itself does not specify?

Well, as this shows, the key isn't set. So, now i can go on making my claim.

In order for the test to run, I had to change the HttpApi*_options
API, to throw an exception type that is defined in ninja, rather than
one defined in HTTP API - the HTTP API already depends on ninja, we
don't need to make the dependency two-way. Patches are coming to
handle that in the API.

Change-Id: Ic2b212a709e2912099ff469f852096a3350ec034
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>