execute_command: Gather comments from working source
We've spent the last two years trying to download downtime ids from an
empty database table. This means that trigger_id has been broken for
two years (though it's inconceivable that someone would *want* to use
it through the GUI), and that service deletion apparently cares not at
all about the comment IDs.

Oh well. This code is destined for a rewrite in the not-too distant future.

Change-Id: Icb69b1a6e347fd8a94abeb994bf47002fae81d91
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>