status_reports: Handle downtime depth properly
When looking for initial downtime depth for services, and we found one
for its host, we completely failed to make use of that. Now, the
initial downtime depth for services is 0, 1 or 2, where 2 means "it
would appear both the host and the service is in downtime". Neither of
these numbers are actually the truth - they're estimates - so they're
cast to "an API bool" (i.e. 0 or 1) for the API's use.

Change-Id: I0a587837d766cb1162a7c1ec05ad493afa2800aa
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>