API break: remove in_scheduled_downtime
So, obviously, bad.

However.

I don't think very many people are using the old API, and if they are,
looking at the "in scheduled downtime" flag is actively misleading. If
anyone finds out the API broke, we can explain to them what happened,
how to proceed, and that their tool actually didn't work in the first
place.

Change-Id: Iff4c33e589e07fc82fe21a0f0a1f3d288dfca65b
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>