old_timeperiod, 7.0 only: Apply ungodly hack due to state in status model
Because we *always* resolve a timeperiod, even when we don't want to
use it, and we do it in the constructor, when we don't yet know if we
are going to use it.

So, let's "break" the validation, by always setting a NULL end_time
to "now". This is the value it gets from httpstatus options, it is
unlikely to interfere with any actual bad values (such as 0, false,
"", or just small numbers), so it should be alright, until the state
removal patches (which broke synergy, so I didn't cherry-pick them)
exists (in 7.1), so this can be reverted.

Change-Id: I70acb06be6661723aa6165768cb3a3ba03c4185e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>