Make database selection work with php-5.5.9 and MySQL 5.6
I have no idea why it doesn't work without this change or
why it works with it, but this patch fixes database selection
on CentOS 7 and recent Debian/Ubuntu systems. Without this
patch, they end up with a MySQL exception about "no database
selected" when trying to run a query.

Change-Id: I1247aaaef1ac943c32da98f6cda797c90e76f2ec
Signed-off-by: Andreas Ericsson <ae@op5.se>