Update favicon.ico icon and fix its links
According to http://en.wikipedia.org/wiki/Favicon, the type
of a favicon.ico in the MS ICO format should be "image/x-icon",
so let's update our links to that. Assuming that actually was
the issue (which it may have been with some versions of
internet explorer), this could actually go a long way towards
fixing issue #9325.

While at it, we also update the logo to a slightly larger
version without the white background, because the previous
one looked completely awful, and especially on high resolution
screens.

Since the icon is no longer a 16x16 one, but rather a 48x48
one, we no longer put it in the 16x16 directory, but instead
put it directly under "icons/", as we don't have a 48x48
directory.

Change-Id: I1308022c7885bcf6313b516f302517ab990732ce
Signed-off-by: Andreas Ericsson <ae@op5.se>