Add support for custom object parents in ORM
To add functionality to several types of objects in ORM, class
inheritance is often useful.

But since we have the structure for example Host_Model =>
BaseHost_Model => Object_Model => BaseObject_Model, we want to inject
the category-class between BaseHost_Model and Object_Model.
BaseHost_Model is autogenereated and can't be changed, so support for
custom object parents is needed in the generator.

In this case, we add a custom parent to all NaemoObjects, which comes
from livestatus, so we can group the functionality of sending commands
to Naemon.

Change-Id: If8b6d8ef5030f65a353d5fc2e445a593174744a0
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>