Remove database tables no longer in use
The removed tables are roles, users, user_roles, user_tokens
and ninja_user_authorization, which are no longer in use.

This is at least the low-hanging fruit of fixing issue #7594.
It's possible there are more database tables that we can get
rid of, but I didn't find any in the 30 or so minutes I spent
looking and I don't think it's worth spending more time on
this.

Change-Id: I03de9b77a7f3998ed5986b3c6aab88bbee78c13d
Signed-off-by: Andreas Ericsson <ae@op5.se>