template/notifications: Escape and format Notify.message() calls
If you add a global notification in a Ninja controller like this:
	$this->add_global_notification("<a href='hello'>hello</a>");

invalid output would be rendered, since we did not escape the ' and
prematurely ended the containing string and thus the first parameter to
Notify.message().

I also added some formatting to make the generated code easier to read.

Change-Id: I32c0a4205d26fd1c409be23c57d6d57eecb2720f
Signed-off-by: Carl Helmertz <chelmertz@op5.com>